+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Update solution to .NET 5.0
 - Tolerate malformed graphics configuration files, binding presets
+- Adjust hyperspace particle flow on ChromaLink devices
 
 ## [1.12.0](https://github.com/poveden/EliteChroma/compare/v1.11.0...v1.12.0) — 2021-05-24